renamed auth routes
