+++ src/features/auth/index.ts
@@ -1,3 +1,3 @@
 export * from './api/register';
 
-export * from './routes';
+export * from './routes/AuthRoutes';
