+++ src/features/auth/routes/index.ts
@@ -1,0 +1,1 @@
+export * from './AuthRoutes';
